+++ app/controllers/v1/paragraphs_controller.rb
@@ -1,5 +1,6 @@
-class V1::ParagraphsController < ApplicationController
+class V1::ParagraphsController < V1::BaseController
   before_action :set_paragraph, only: [:show, :update, :previous_version, :destroy, :accept]
+  before_action :authenticate_user!, except: [:index, :show, :previous_version]
 
   def index
     @paragraphs = Paragraph.where(document_id: params[:document_id], 'suggested_to' => nil)
@@ -16,7 +17,6 @@ class V1::ParagraphsController < ApplicationController
 
   def create
     @paragraph = Paragraph.new paragraph_params
-    @paragraph.document_id = params[:document_id]
     if @paragraph.save
       render json: @paragraph
     else
@@ -25,11 +25,13 @@ class V1::ParagraphsController < ApplicationController
   end
 
   def update
+    authorize @paragraph
     @paragraph.update_attributes paragraph_params
     render json: @paragraph
   end
 
   def accept
+    authorize @paragraph
     if @paragraph.accept
       render json: {}, status: 200
     else
@@ -38,6 +40,7 @@ class V1::ParagraphsController < ApplicationController
   end
 
   def destroy
+    authorize @paragraph
     if @paragraph.destroy
       render json: {}, status: 200
     else
@@ -46,12 +49,11 @@ class V1::ParagraphsController < ApplicationController
   end
 
   private
-
     def set_paragraph
       @paragraph = Paragraph.find(params[:id])
     end
 
     def paragraph_params
-      params.require(:paragraph).permit(:description, :suggested_to)
+      params.require(:paragraph).permit(:description, :suggested_to, :document_id)
     end
 end
